Removed constructor from TickeInfo New objects should be created through a GUI, plus update code and constructer code were almost exactly the same GUI creates a new object that gets its symbol set, then gets updated with current info then passed back to the main app, ready to be called to be displayed Because whenever you create an object, you also need to update it too.
What I needed to do was make a setter in TickerInfo to set the symbol first
then the update method could parse the value out of the object and update
all of that objects variables, and return the finished object back to the
main method. Simple, right?

We are starting to get away from leaning and more into the creative side
of things.

Also
updated Javadocs, yay.

... I have no idea what I'm doing, shhhh, don't tell anyone.